Swamp_Mushrooms/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
paho-mqtt==1.6.1
pydantic==1.10.11
numpy==1.26.4
scipy==1.11.3
python-dotenv==1.0.0
sqlitedict==2.0.0
